Implement forward heart beat to servers

+++ naming/src/main/java/com/alibaba/nacos/naming/remote/RemotingConnectionHolder.java
@@ -16,21 +16,29 @@
 
 package com.alibaba.nacos.naming.remote;
 
+import com.alibaba.nacos.api.common.Constants;
+import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.api.remote.connection.Connection;
 import com.alibaba.nacos.common.notify.Event;
 import com.alibaba.nacos.common.notify.NotifyCenter;
 import com.alibaba.nacos.common.notify.listener.Subscriber;
 import com.alibaba.nacos.core.remote.ClientConnectionEventListener;
 import com.alibaba.nacos.core.remote.event.RemotingHeartBeatEvent;
+import com.alibaba.nacos.naming.cluster.remote.ClusterClient;
+import com.alibaba.nacos.naming.cluster.remote.ClusterClientManager;
+import com.alibaba.nacos.naming.cluster.remote.request.ForwardHeartBeatRequest;
 import com.alibaba.nacos.naming.core.ServiceManager;
+import com.alibaba.nacos.naming.misc.GlobalExecutor;
 import com.alibaba.nacos.naming.misc.Loggers;
 import com.alibaba.nacos.naming.push.RemotePushService;
 import com.alibaba.nacos.naming.remote.task.RenewInstanceBeatTask;
 import com.alibaba.nacos.naming.remote.worker.RemotingWorkersManager;
 import org.springframework.stereotype.Component;
 
+import java.util.Collection;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Remoting connection holder.
@@ -46,10 +54,13 @@ public class RemotingConnectionHolder extends ClientConnectionEventListener {
     
     private final ServiceManager serviceManager;
     
-    public RemotingConnectionHolder(RemotePushService remotePushService, ServiceManager serviceManager) {
+    public RemotingConnectionHolder(RemotePushService remotePushService, ServiceManager serviceManager,
+            ClusterClientManager clusterClientManager) {
         this.remotePushService = remotePushService;
         this.serviceManager = serviceManager;
-        NotifyCenter.registerSubscriber(new RemotingHeartBeatSubscriber(this));
+        NotifyCenter.registerSubscriber(new RemotingHeartBeatSubscriber(this, clusterClientManager));
+        GlobalExecutor.scheduleRemoteConnectionManager(new RemotingConnectionCleaner(this), 0,
+                Constants.DEFAULT_HEART_BEAT_INTERVAL, TimeUnit.MILLISECONDS);
     }
     
     @Override
@@ -62,9 +73,15 @@ public class RemotingConnectionHolder extends ClientConnectionEventListener {
     
     @Override
     public void clientDisConnected(Connection connect) {
-        Loggers.SRV_LOG
-                .info("Client connection {} disconnect, remove instances and subscribers", connect.getConnectionId());
-        RemotingConnection remotingConnection = connectionCache.remove(connect.getConnectionId());
+        clientDisConnected(connect.getConnectionId());
+    }
+    
+    private void clientDisConnected(String connectionId) {
+        Loggers.SRV_LOG.info("Client connection {} disconnect, remove instances and subscribers", connectionId);
+        RemotingConnection remotingConnection = connectionCache.remove(connectionId);
+        if (null == remotingConnection) {
+            return;
+        }
         for (String each : remotingConnection.getSubscriberIndex().keySet()) {
             remotePushService.removeAllSubscribeForService(each);
         }
@@ -72,6 +89,10 @@ public class RemotingConnectionHolder extends ClientConnectionEventListener {
     
     public RemotingConnection getRemotingConnection(String connectionId) {
         return connectionCache.get(connectionId);
+    }
+    
+    public Collection<String> getAllConnectionId() {
+        return connectionCache.keySet();
     }
     
     /**
@@ -84,25 +105,61 @@ public class RemotingConnectionHolder extends ClientConnectionEventListener {
             return;
         }
         RemotingConnection remotingConnection = connectionCache.get(connectionId);
+        remotingConnection.setLastHeartBeatTime(System.currentTimeMillis());
         RemotingWorkersManager.dispatch(connectionId, new RenewInstanceBeatTask(remotingConnection, serviceManager));
     }
     
     private static class RemotingHeartBeatSubscriber extends Subscriber<RemotingHeartBeatEvent> {
         
         private final RemotingConnectionHolder remotingConnectionHolder;
-    
-        public RemotingHeartBeatSubscriber(RemotingConnectionHolder remotingConnectionHolder) {
+        
+        private final ClusterClientManager clusterClientManager;
+        
+        public RemotingHeartBeatSubscriber(RemotingConnectionHolder remotingConnectionHolder,
+                ClusterClientManager clusterClientManager) {
             this.remotingConnectionHolder = remotingConnectionHolder;
+            this.clusterClientManager = clusterClientManager;
         }
-    
+        
         @Override
         public void onEvent(RemotingHeartBeatEvent event) {
             remotingConnectionHolder.renewRemotingConnection(event.getConnectionId());
+            for (ClusterClient each : clusterClientManager.getAllClusterClient()) {
+                try {
+                    each.request(new ForwardHeartBeatRequest(event.getConnectionId()));
+                } catch (NacosException nacosException) {
+                    Loggers.DISTRO.warn("Forward heart beat failed.", nacosException);
+                }
+            }
         }
-    
+        
         @Override
         public Class<? extends Event> subscribeType() {
             return RemotingHeartBeatEvent.class;
+        }
+    }
+    
+    private static class RemotingConnectionCleaner implements Runnable {
+        
+        private final RemotingConnectionHolder remotingConnectionHolder;
+        
+        public RemotingConnectionCleaner(RemotingConnectionHolder remotingConnectionHolder) {
+            this.remotingConnectionHolder = remotingConnectionHolder;
+        }
+        
+        @Override
+        public void run() {
+            long currentTime = System.currentTimeMillis();
+            for (String each : remotingConnectionHolder.getAllConnectionId()) {
+                RemotingConnection remotingConnection = remotingConnectionHolder.getRemotingConnection(each);
+                if (null != remotingConnection && isExpireConnection(currentTime, remotingConnection)) {
+                    remotingConnectionHolder.clientDisConnected(each);
+                }
+            }
+        }
+        
+        private boolean isExpireConnection(long currentTime, RemotingConnection remotingConnection) {
+            return remotingConnection.getLastHeartBeatTime() - currentTime > Constants.DEFAULT_IP_DELETE_TIMEOUT * 2;
         }
     }
 }
